+++ database/seeders/DatabaseSeeder.php
@@ -2,8 +2,10 @@
 
 namespace Database\Seeders;
 
+use App\Models\Product;
+use App\Models\Review;
 use App\Models\User;
-// use Illuminate\Database\Console\Seeds\WithoutModelEvents;
+use Database\Factories\UserFactory;
 use Illuminate\Database\Seeder;
 
 class DatabaseSeeder extends Seeder
@@ -13,11 +15,27 @@ class DatabaseSeeder extends Seeder
      */
     public function run(): void
     {
-        // User::factory(10)->create();
+        $users =  User::factory(5)->create();
+        $newCredentials = UserFactory::getCredentials();
 
-        User::factory()->create([
-            'name' => 'Test User',
-            'email' => 'test@example.com'
-        ]);
+        $filepath = storage_path("app/public/users/users_data.json");
+        $existingData = [];
+        if (file_exists($filepath)) {
+            $jsonContent = file_get_contents($filepath);
+            $existingData = json_decode($jsonContent, true) ?? [];
+        }
+        $newUsersData = array_merge($existingData, $newCredentials);
+        file_put_contents($filepath, json_encode($newUsersData, JSON_PRETTY_PRINT));
+
+        $products = Product::factory(5)->make()->each(function ($product) use ($users) {
+            $product->user_id = $users->random()->id;
+            $product->save();
+        });
+
+        Review::factory(5)->make()->each(function ($review) use ($users, $products) {
+            $review->user_id = $users->random()->id;
+            $review->product_id = $products->random()->id;
+            $review->save();
+        });
     }
 }